refactor json utils
