+++ src/components/ui/code.tsx
@@ -3,6 +3,7 @@ import { Button } from "@/src/components/ui/button";
 import { Check, ChevronsDownUp, ChevronsUpDown, Copy } from "lucide-react";
 import { cn } from "@/src/utils/tailwind";
 import { default as React18JsonView } from "react18-json-view";
+import { deepParseJson, parseJson } from "@/src/utils/json";
 
 export function JSONView(props: {
   json?: unknown;
@@ -105,49 +106,4 @@ export function CodeView(props: {
       </div>
     </div>
   );
-}
-
-const parseJson = (input: string) => {
-  try {
-    return JSON.parse(input) as unknown;
-  } catch {
-    return input;
-  }
-};
-
-/**
- * Deeply parses a JSON string or object for nested stringified JSON
- * @param json JSON string or object to parse
- * @returns Parsed JSON object
- */
-function deepParseJson(json: unknown): unknown {
-  if (typeof json === "string") {
-    try {
-      // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-      const parsed = JSON.parse(json);
-      return deepParseJson(parsed); // Recursively parse parsed value
-    } catch (e) {
-      return json; // If it's not a valid JSON string, just return the original string
-    }
-  } else if (typeof json === "object" && json !== null) {
-    // Handle arrays
-    if (Array.isArray(json)) {
-      for (let i = 0; i < json.length; i++) {
-        json[i] = deepParseJson(json[i]);
-      }
-    } else {
-      // Handle nested objects
-      for (const key in json) {
-        // Ensure we only iterate over the object's own properties
-        if (Object.prototype.hasOwnProperty.call(json, key)) {
-          (json as Record<string, unknown>)[key] = deepParseJson(
-            (json as Record<string, unknown>)[key],
-          );
-        }
-      }
-    }
-    return json;
-  }
-
-  return json;
 }
